test(trusty-memory): give the fd-bound assertion its own test process
CI caught a real flaw in the #4639 fd test: `fd_metrics::count_open_fds`
reports the WHOLE process, but cargo runs a target's tests as parallel
threads of one process. Sitting alongside the crate's ~480 other lib tests,
the measurement was swamped by their unrelated fd churn — CI saw a +811
delta against a 120-fd signal and failed. My local run had filtered to
`--lib session_store_cache`, which ran only 7 tests and hid the contention.

Cargo gives each `tests/*.rs` file its own process, so move the assertion to
`tests/session_store_fd_bound.rs` as the sole test in that target, where the
delta is attributable. The file header says why nothing else may be added
to it.

The assertion is unchanged and still has teeth: with the cap removed it
fails with "grew by 120 (before=3, after=123)" — exactly one leaked fd per
palace. The fd check now runs before the entry-count check so the primary
claim is what reports.
